🚚 [packages] Move class `PackageRepository` to module `repository`
