+++ src/cutty/packages/domain/package.py
@@ -1,12 +1,9 @@
 """Package."""
 from __future__ import annotations
 
-import abc
-from collections.abc import Iterator
 from dataclasses import dataclass
 from typing import Optional
 
-from cutty.compat.contextlib import contextmanager
 from cutty.filesystems.domain.path import Path
 from cutty.filesystems.domain.pathfs import PathFilesystem
 from cutty.filesystems.domain.purepath import PurePath
@@ -29,12 +26,3 @@ class Package:
             Path(filesystem=PathFilesystem(path)),
             self.revision,
         )
-
-
-class PackageRepository(abc.ABC):
-    """A package repository."""
-
-    @contextmanager
-    @abc.abstractmethod
-    def get(self, revision: Optional[Revision] = None) -> Iterator[Package]:
-        """Retrieve the package with the given revision."""
