+++ src/cutty/packages/domain/repository.py
@@ -1,0 +1,17 @@
+"""Package repositories."""
+import abc
+from collections.abc import Iterator
+from typing import Optional
+
+from cutty.compat.contextlib import contextmanager
+from cutty.packages.domain.package import Package
+from cutty.packages.domain.revisions import Revision
+
+
+class PackageRepository(abc.ABC):
+    """A package repository."""
+
+    @contextmanager
+    @abc.abstractmethod
+    def get(self, revision: Optional[Revision] = None) -> Iterator[Package]:
+        """Retrieve the package with the given revision."""
