+++ tests/fixtures/packages/domain/providers.py
@@ -10,8 +10,8 @@ from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path
 from cutty.packages.domain.locations import Location
 from cutty.packages.domain.package import Package
-from cutty.packages.domain.package import PackageRepository
 from cutty.packages.domain.providers import Provider
+from cutty.packages.domain.repository import PackageRepository
 from cutty.packages.domain.revisions import Revision
